收货地址接口修改

+++ src/main/java/com/jichuangtech/clothshopserver/constant/UserAddressConstant.java
@@ -3,6 +3,6 @@ package com.jichuangtech.clothshopserver.constant;
 public class UserAddressConstant extends Constant{
 	public static final String API_USER_ADDRESS = API + "/useraddress";
 	
-	public static final int DEFAULT_ADDRESS = 0;
-	public static final int NOT_DEFAULT_ADDRESS = 1;
+	public static final byte DEFAULT_ADDRESS = 0;
+	public static final byte NOT_DEFAULT_ADDRESS = 1;
 }
